Move campaign page into core
