+++ core/templates/core/campaign.html
