--- marketing/views.py
@@ -1,14 +1,0 @@
-from django.views.generic import TemplateView
-
-from core.mixins import GetCMSPageMixin
-
-
-class CampaignPageView(
-    GetCMSPageMixin,
-    TemplateView
-):
-    template_name = 'marketing/campaign.html'
-
-    @property
-    def slug(self):
-        return self.kwargs['slug']
